+++ src/backends/sentry_backend_crashpad.cpp
@@ -95,7 +95,7 @@ sentry__crashpad_backend_shutdown(sentry_backend_t *backend)
     data->db = nullptr;
 }
 
-static void
+static bool
 sentry__crashpad_backend_startup(
     sentry_backend_t *backend, const sentry_options_t *options)
 {
@@ -128,9 +128,9 @@ sentry__crashpad_backend_startup(
     sentry__path_free(owned_handler_path);
     if (!absolute_handler_path
         || !sentry__path_is_file(absolute_handler_path)) {
-        SENTRY_DEBUG("unable to start crashpad backend, invalid handler_path");
+        SENTRY_WARN("unable to start crashpad backend, invalid handler_path");
         sentry__path_free(absolute_handler_path);
-        return;
+        return false;
     }
 
     SENTRY_TRACEF("starting crashpad backend with handler "
@@ -195,8 +195,11 @@ sentry__crashpad_backend_startup(
     if (success) {
         SENTRY_DEBUG("started crashpad client handler");
     } else {
-        SENTRY_DEBUG("failed to start crashpad client handler");
-        return;
+        SENTRY_WARN("failed to start crashpad client handler");
+        // not calling `shutdown`
+        delete data->db;
+        data->db = nullptr;
+        return false;
     }
 
     if (!options->system_crash_reporter_enabled) {
@@ -207,6 +210,7 @@ sentry__crashpad_backend_startup(
         crashpad_info->set_system_crash_reporter_forwarding(
             crashpad::TriState::kDisabled);
     }
+    return true;
 }
 
 static void