+++ src/sentry_transport.h
@@ -25,7 +25,7 @@ void sentry__transport_send_envelope(
 /**
  * Calls the transports startup hook.
  */
-void sentry__transport_startup(
+bool sentry__transport_startup(
     sentry_transport_t *transport, const sentry_options_t *options);
 
 /**